refactor: move blacklist functions to their own file
