+++ lister/tmux.go
@@ -2,30 +2,12 @@ package lister
 
 import (
 	"fmt"
-	"strings"
 
 	"github.com/joshmedeski/sesh/model"
 )
 
 func tmuxKey(name string) string {
 	return fmt.Sprintf("tmux:%s", name)
-}
-
-func isBlacklisted(blacklist []string, name string) bool {
-	for _, blacklistedName := range blacklist {
-		if strings.EqualFold(blacklistedName, name) {
-			return true
-		}
-	}
-	return false
-}
-
-func createBlacklistSet(blacklist []string) map[string]struct{} {
-	blacklistSet := make(map[string]struct{}, len(blacklist))
-	for _, name := range blacklist {
-		blacklistSet[name] = struct{}{}
-	}
-	return blacklistSet
 }
 
 func listTmux(l *RealLister) (model.SeshSessions, error) {
